Expose commit in externalities.
Also refactor `TestExternalities` into a single place.
